1379 | Move tests to separate files
