+++ tests/i3c_ccc_tests.py
@@ -181,6 +181,111 @@ class TestSupernovaController(unittest.TestCase):
 
         # Idem (BMC2-1663)
         # print(self.i3c.ccc_getxtime(0x08)) 
+        
+    def test_ccc_getpid(self):
+        if not self.use_simulator:
+            self.skipTest("For simulator only")
+
+        i3c = self.device.create_interface("i3c.controller")
+
+        i3c.init_bus(3300)
+
+        (success, result) = i3c.ccc_getpid(0x08)
+
+        self.assertTupleEqual((success, result), (True, [0x00, 0x00, 0x00, 0x00, 0x64, 0x65]))
+
+    def test_ccc_rstdaa(self):
+        if self.use_simulator:
+            self.skipTest("For real device only")
+
+        i3c = self.device.create_interface("i3c.controller")
+
+        i3c.init_bus(3300)
+
+        (success, result) = i3c.ccc_rstdaa()
+
+        self.assertTupleEqual((success, result), (True, None))
+
+        (success, result) = i3c.ccc_getpid(0x08)
+
+        self.assertTupleEqual((success, result), (False, "NACK_ERROR"))
+
+    def test_ccc_get_and_set_mrl(self):
+        if not self.use_simulator:
+            self.skipTest("For simulator only")
+
+        i3c = self.device.create_interface("i3c.controller")
+
+        i3c.init_bus(3300)
+
+        (success, result) = i3c.ccc_getmrl(0x08)
+
+        self.assertTupleEqual((success, result), (True, 16))
+
+        (success, _) = i3c.ccc_unicast_setmrl(0x08, 10)
+
+        self.assertEqual(success, True)
+
+        (success, result) = i3c.ccc_getmrl(0x08)
+
+        self.assertTupleEqual((success, result), (True, 10))
+
+    def test_ccc_set_mwl_real_device(self):
+        if self.use_simulator:
+            self.skipTest("For real device only")
+
+        i3c = self.device.create_interface("i3c.controller")
+
+        i3c.init_bus(3300)
+
+        (_, connectedDevices) = i3c.targets()
+
+        if len(connectedDevices) == 0:
+            self.fail("Test requires a connected target")
+
+        (success, _) = i3c.ccc_broadcast_setmwl(10)
+
+        self.assertEqual(success, True)
+
+        (success, result) = i3c.ccc_getmwl(0x08)
+
+        self.assertTupleEqual((success, result), (True, 10))
+
+    def test_direct_rstact_read(self):
+        if self.use_simulator:
+            self.skipTest("For real device only")
+
+        i3c = self.device.create_interface("i3c.controller")
+
+        i3c.init_bus(3300)
+
+        (success, result) = i3c.ccc_direct_rstact(0x08,I3cTargetResetDefByte.RESET_I3C_PERIPHERAL, TransferDirection.READ)
+
+        self.assertTupleEqual((success, result), (True, [0x00]))
+
+    def test_direct_rstact_write(self):
+        if self.use_simulator:
+            self.skipTest("For real device only")
+
+        i3c = self.device.create_interface("i3c.controller")
+
+        i3c.init_bus(3300)
+
+        (success, result) = i3c.ccc_direct_rstact(0x08,I3cTargetResetDefByte.RESET_I3C_PERIPHERAL, TransferDirection.WRITE)
+
+        self.assertTupleEqual((success, result), (True, None))
+
+    def test_broadcast_rstact(self):
+        if self.use_simulator:
+            self.skipTest("For real device only")
+
+        i3c = self.device.create_interface("i3c.controller")
+
+        i3c.init_bus(3300)
+
+        (success, result) = i3c.ccc_broadcast_rstact(I3cTargetResetDefByte.RESET_I3C_PERIPHERAL)
+
+        self.assertTupleEqual((success, result), (True, None))
 
 if __name__ == "__main__":
     unittest.main()
