+++ tests/test.py
@@ -16,30 +16,10 @@ from BinhoSupernova.commands.definitions import (
     SpiControllerBitOrder, SpiControllerMode, SpiControllerDataWidth,
     SpiControllerChipSelect, SpiControllerChipSelectPolarity,
     GpioPinNumber, GpioLogicLevel, GpioFunctionality, GpioTriggerType)
-from BinhoSupernova.Supernova import I3cTargetResetDefByte
-from BinhoSupernova.Supernova import TransferDirection
 
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
 from binhosimulators import BinhoSupernovaSimulator
 
-counter = 0
-last_ibi = Event()
-caught_ibis = []
-
-def _is_ibi(name, message):
-        return message["name"].strip() == "I3C IBI NOTIFICATION" and message["header"]["type"] == "IBI_NORMAL"
-
-def _handle_ibi(name, message):
-    global counter
-    global last_ibi
-    global caught_ibis
-
-    caught_ibis.append({"dynamic_address": message["header"]["address"],  "controller_response": message["header"]["response"], "mdb":message["payload"][0]})
-
-    counter += 1
-    if counter == 5:
-        last_ibi.set()
-        
 class TestSupernovaController(unittest.TestCase):
     @classmethod
     def setUpClass(cls):
@@ -58,13 +38,6 @@ class TestSupernovaController(unittest.TestCase):
         self.device_info = self.device.open()
 
     def tearDown(self):
-        global caught_ibis
-        global last_ibi
-        global counter
-
-        counter = 0
-        caught_ibis.clear()
-        last_ibi.clear()
         self.device.close()
 
     def test_open_device_with_wrong_address(self):
@@ -102,176 +75,6 @@ class TestSupernovaController(unittest.TestCase):
         with self.assertRaises(UnknownInterfaceError):
             self.device.create_interface("foo")
 
-    def test_set_i3c_bus_voltage_attribute(self):
-        i3c = self.device.create_interface("i3c.controller")
-
-        (success, result) = i3c.set_bus_voltage(3300)
-
-        self.assertTupleEqual((success, result), (True, 3300))
-        self.assertEqual(i3c.bus_voltage, 3300)
-
-    def test_set_i3c_frequencies(self):
-        i3c = self.device.create_interface("i3c.controller")
-
-        (success, result) = i3c.set_parameters(
-            i3c.I3cPushPullTransferRate.PUSH_PULL_3_75_MHZ,
-            i3c.I3cOpenDrainTransferRate.OPEN_DRAIN_100_KHZ
-        )
-
-        self.assertEqual((success, result), (True, (i3c.I3cPushPullTransferRate.PUSH_PULL_3_75_MHZ, i3c.I3cOpenDrainTransferRate.OPEN_DRAIN_100_KHZ)))
-
-    def test_i3c_frequencies_should_take_default_values_if_not_set(self):
-        i3c = self.device.create_interface("i3c.controller")
-
-        (success, result) = i3c.get_parameters()
-
-        self.assertEqual((success, result), (True, (i3c.I3cPushPullTransferRate.PUSH_PULL_3_75_MHZ, i3c.I3cOpenDrainTransferRate.OPEN_DRAIN_100_KHZ)))
-
-    def test_set_i3c_bus_voltage_attribute_error(self):
-        if not self.use_simulator:
-            self.skipTest("For simulator only")
-
-        i3c = self.device.create_interface("i3c.controller")
-
-        # Mock the controller's sync_submit method to simulate a failure response.
-        # It's important to note that this approach makes the test somewhat dependent
-        # on the internal implementation of the device class.
-        with mock.patch.object(self.device.controller, "sync_submit") as mock_sync_submit:
-            mock_sync_submit.return_value = [{"name": "SET I3C BUS VOLTAGE", "result": 1}]  # Simulate an error response
-
-            # Call the method under test
-            (success, result) = i3c.set_bus_voltage(3300)
-
-            # Assert that the method correctly returns an error
-            self.assertTupleEqual((success, result), (False, "Set bus voltage failed"))
-            self.assertEqual(i3c.bus_voltage, None)
-
-    def test_i3c_init_bus_with_no_targets_connected(self):
-        # This test assumes that the are no devices connected to the bus.
-
-        if self.use_simulator:
-            self.skipTest("For real device only")
-
-        i3c = self.device.create_interface("i3c.controller")
-
-        (success, result) = i3c.init_bus(1500)
-
-        self.assertEqual(success, False, "Init bus succeded, are there devices connected?")
-        self.assertTrue("errors" in result)
-
-    def test_i3c_reset_bus(self):
-        if not self.use_simulator:
-            self.skipTest("For simulator only")
-
-        i3c = self.device.create_interface("i3c.controller")
-
-        i3c.init_bus(2000)
-        (success, result) = i3c.reset_bus()
-
-        self.assertTupleEqual((success, result), (True, 2000))
-
-    def test_i3c_reset_bus_before_init(self):
-        if not self.use_simulator:
-            self.skipTest("For simulator only")
-
-        i3c = self.device.create_interface("i3c.controller")
-
-        (success, _) = i3c.reset_bus()
-
-        self.assertEqual(success, True)
-
-    def test_i3c_targets(self):
-        if not self.use_simulator:
-            self.skipTest("For simulator only")
-
-        i3c = self.device.create_interface("i3c.controller")
-
-        i3c.init_bus(3300)
-        (success, targets) = i3c.targets()
-
-        self.assertEqual(success, True)
-        self.assertEqual(len(targets), 3)
-        self.assertDictEqual(targets[0], {
-            "static_address": 0x50,
-            "dynamic_address": 0x08,
-            "bcr": 0x10,
-            "dcr": 0xC3,
-            "pid": ["0x00", "0x00", "0x00", "0x00", "0x64", "0x65"],
-        })
-        self.assertDictEqual(targets[1], {
-            "static_address": 0x51,
-            "dynamic_address": 0x09,
-            "bcr": 0x03,
-            "dcr": 0x63,
-            "pid": ["0x02", "0x17", "0x0F", "0x1D", "0x00", "0x5A"],
-        })
-        self.assertDictEqual(targets[2], {
-            "static_address": 0x52,
-            "dynamic_address": 0x0A,
-            "bcr": 0x10,
-            "dcr": 0xC3,
-            "pid": ["0x06", "0x06", "0x06", "0x06", "0x66", "0x66"],
-        })
-
-    def test_i3c_targets_when_bus_not_initialized(self):
-        if not self.use_simulator:
-            self.skipTest("For simulator only")
-
-        i3c = self.device.create_interface("i3c.controller")
-
-        (success, targets) = i3c.targets()
-
-        self.assertTupleEqual((success, targets), (True, []))
-
-    def test_i3c_write_operation_when_bus_not_initialized(self):
-        if not self.use_simulator:
-            self.skipTest("For simulator only")
-
-        i3c = self.device.create_interface("i3c.controller")
-
-        (success, _) = i3c.write(
-            0x08,
-            i3c.TransferMode.I3C_SDR,
-            [0x00, 0x00],
-            [0xDE, 0xAD, 0xBE, 0xEF]
-        )
-
-        self.assertEqual(success, False)
-
-    def test_i3c_read_operation_when_bus_not_initialized(self):
-        if not self.use_simulator:
-            self.skipTest("For simulator only")
-
-        i3c = self.device.create_interface("i3c.controller")
-
-        (success, _) = i3c.read(
-            0x08,
-            i3c.TransferMode.I3C_SDR,
-            [0x00, 0x00],
-            1
-        )
-
-        self.assertEqual(success, False)
-
-    def test_i3c_successful_write_operation_on_target_should_return_none(self):
-        if not self.use_simulator:
-            self.skipTest("For simulator only")
-
-        i3c = self.device.create_interface("i3c.controller")
-
-        i3c.init_bus(3300)
-
-        subaddress = [0x00, 0x00]
-
-        (success, result) = i3c.write(
-            0x08,
-            i3c.TransferMode.I3C_SDR,
-            subaddress,
-            [0xDE, 0xAD, 0xBE, 0xEF]
-        )
-
-        self.assertTupleEqual((success, result), (True, None))
-
     # Mock the controller's sync_submit method to simulate an exception.
     # It's important to note that this approach makes the test somewhat dependent
     # on the internal implementation of the device class.
@@ -282,175 +85,6 @@ class TestSupernovaController(unittest.TestCase):
 
         with self.assertRaises(BackendError):
             self.device.get_hardware_version()
-
-    def test_i3c_successful_write_read_operations_on_target(self):
-        if not self.use_simulator:
-            self.skipTest("For simulator only")
-
-        i3c = self.device.create_interface("i3c.controller")
-
-        i3c.init_bus(3300)
-
-        subaddress = [0x00, 0x00]
-
-        (success, result) = i3c.write(
-            0x08,
-            i3c.TransferMode.I3C_SDR,
-            subaddress,
-            [0xDE, 0xAD, 0xBE, 0xEF]
-        )
-
-        (success, result) = i3c.read(
-            0x08,
-            i3c.TransferMode.I3C_SDR,
-            subaddress,
-            4
-        )
-
-        self.assertTupleEqual((success, result), (True, [0xDE, 0xAD, 0xBE, 0xEF]))
-
-    def test_ccc_getpid(self):
-        if not self.use_simulator:
-            self.skipTest("For simulator only")
-
-        i3c = self.device.create_interface("i3c.controller")
-
-        i3c.init_bus(3300)
-
-        (success, result) = i3c.ccc_getpid(0x08)
-
-        self.assertTupleEqual((success, result), (True, [0x00, 0x00, 0x00, 0x00, 0x64, 0x65]))
-
-    def test_ccc_rstdaa(self):
-        if self.use_simulator:
-            self.skipTest("For real device only")
-
-        i3c = self.device.create_interface("i3c.controller")
-
-        i3c.init_bus(3300)
-
-        (success, result) = i3c.ccc_rstdaa()
-
-        self.assertTupleEqual((success, result), (True, None))
-
-        (success, result) = i3c.ccc_getpid(0x08)
-
-        self.assertTupleEqual((success, result), (False, "NACK_ERROR"))
-
-    def test_ccc_get_and_set_mrl(self):
-        if not self.use_simulator:
-            self.skipTest("For simulator only")
-
-        i3c = self.device.create_interface("i3c.controller")
-
-        i3c.init_bus(3300)
-
-        (success, result) = i3c.ccc_getmrl(0x08)
-
-        self.assertTupleEqual((success, result), (True, 16))
-
-        (success, _) = i3c.ccc_unicast_setmrl(0x08, 10)
-
-        self.assertEqual(success, True)
-
-        (success, result) = i3c.ccc_getmrl(0x08)
-
-        self.assertTupleEqual((success, result), (True, 10))
-
-    def test_ccc_set_mwl_real_device(self):
-        if self.use_simulator:
-            self.skipTest("For real device only")
-
-        i3c = self.device.create_interface("i3c.controller")
-
-        i3c.init_bus(3300)
-
-        (_, connectedDevices) = i3c.targets()
-
-        if len(connectedDevices) == 0:
-            self.fail("Test requires a connected target")
-
-        (success, _) = i3c.ccc_broadcast_setmwl(10)
-
-        self.assertEqual(success, True)
-
-        (success, result) = i3c.ccc_getmwl(0x08)
-
-        self.assertTupleEqual((success, result), (True, 10))
-
-    def test_toggle_handle_ibi(self):
-        if not self.use_simulator:
-            self.skipTest("For simulator only")
-        
-        self.device.on_notification(name="ibi", filter_func=_is_ibi, handler_func=_handle_ibi)
-
-        i3c = self.device.create_interface("i3c.controller")
-
-        i3c.init_bus(3300)
-
-        (success, _) = i3c.toggle_ibi(0x0A, True)
-
-        self.assertEqual(success, True)
-
-        last_ibi.wait()
-
-        (success, _) = i3c.toggle_ibi(0x0A, False)
-
-        self.assertEqual(success, True)
-
-        self.assertEqual(len(caught_ibis), 5)
-        for ibi in caught_ibis:
-            self.assertDictEqual({"dynamic_address": 10, "controller_response": "IBI_ACKED_WITH_PAYLOAD", "mdb": 2}, ibi)
-
-    def test_find_target_by_pid_with_BMI323(self):
-        if self.use_simulator:
-            self.skipTest("For real device only")
-
-        i3c = self.device.create_interface("i3c.controller")
-        i3c.init_bus(3300)
-
-        target_pid = ["0x07", "0x70", "0x10", "0x43", "0x10", "0x00"]
-        (deviceFound, bmi_device) = i3c.find_target_device_by_pid(target_pid)
-
-        self.assertEqual(True, deviceFound, "Failed to find BMI323, is it connected?")
-        self.assertDictContainsSubset({"bcr": 6, "dcr": 239, "pid": ["0x07", "0x70", "0x10", "0x43", "0x10", "0x00"]}, bmi_device, "Failed to find BMI323, is it connected?")
-
-    def test_toggle_handle_ibi_BMI323(self):
-        if self.use_simulator:
-            self.skipTest("For real device only")
-        
-        self.device.on_notification(name="ibi", filter_func=_is_ibi, handler_func=_handle_ibi)
-
-        i3c = self.device.create_interface("i3c.controller")
-        i3c.init_bus(3300)
-
-        target_pid = ["0x07", "0x70", "0x10", "0x43", "0x10", "0x00"]
-        (deviceFound, bmi_device) = i3c.find_target_device_by_pid(target_pid)
-
-        if deviceFound is False:
-            self.skipTest("BMI device not found in the I3C bus")
-
-        target_address = bmi_device["dynamic_address"]
-
-        i3c.toggle_ibi(target_address, False)
-
-        i3c.write(target_address, i3c.TransferMode.I3C_SDR, [0x20], [0x85, 0x40])
-        i3c.write(target_address, i3c.TransferMode.I3C_SDR, [0x21], [0x95, 0x40])
-        i3c.write(target_address, i3c.TransferMode.I3C_SDR, [0x3A], [0x00, 0x00])
-        i3c.write(target_address, i3c.TransferMode.I3C_SDR, [0x3B], [0x00, 0x0C])
-        i3c.write(target_address, i3c.TransferMode.I3C_SDR, [0x38], [0x05, 0x05])
-
-        i3c.toggle_ibi(target_address, True)
-
-        last_ibi.wait()
-
-        (success, _) = i3c.toggle_ibi(target_address, False)
-
-        self.assertEqual(success, True)
-
-        self.assertEqual(len(caught_ibis), 5)
-        for ibi in caught_ibis:
-            self.assertDictEqual({"dynamic_address": target_address, "controller_response": "IBI_ACKED_WITH_PAYLOAD", "mdb": 2}, ibi)
 
     def test_spi_controller_set_bus_voltage(self):
         spi_controller = self.device.create_interface("spi.controller")
@@ -594,54 +228,5 @@ class TestSupernovaController(unittest.TestCase):
         
         (success, result) = gpio.disable_interrupt(GpioPinNumber.GPIO_5)
         self.assertTupleEqual((success, result), (True, None))
-
-    def test_direct_rstact_read(self):
-        if self.use_simulator:
-            self.skipTest("For real device only")
-
-        i3c = self.device.create_interface("i3c.controller")
-
-        i3c.init_bus(3300)
-
-        (success, result) = i3c.ccc_direct_rstact(0x08,I3cTargetResetDefByte.RESET_I3C_PERIPHERAL, TransferDirection.READ)
-
-        self.assertTupleEqual((success, result), (True, [0x00]))
-
-    def test_direct_rstact_write(self):
-        if self.use_simulator:
-            self.skipTest("For real device only")
-
-        i3c = self.device.create_interface("i3c.controller")
-
-        i3c.init_bus(3300)
-
-        (success, result) = i3c.ccc_direct_rstact(0x08,I3cTargetResetDefByte.RESET_I3C_PERIPHERAL, TransferDirection.WRITE)
-
-        self.assertTupleEqual((success, result), (True, None))
-
-    def test_broadcast_rstact(self):
-        if self.use_simulator:
-            self.skipTest("For real device only")
-
-        i3c = self.device.create_interface("i3c.controller")
-
-        i3c.init_bus(3300)
-
-        (success, result) = i3c.ccc_broadcast_rstact(I3cTargetResetDefByte.RESET_I3C_PERIPHERAL)
-
-        self.assertTupleEqual((success, result), (True, None))
-
-    def test_trigger_target_reset_pattern(self):
-        if self.use_simulator:
-            self.skipTest("For real device only")
-
-        i3c = self.device.create_interface("i3c.controller")
-
-        i3c.init_bus(3300)
-
-        (success, result) = i3c.trigger_target_reset_pattern()
-
-        self.assertTupleEqual((success, result), (True, None))
-
 if __name__ == "__main__":
     unittest.main()
